services/orchestrator/pyproject.toml
[tool.poetry]
name = "orchestrator"
version = "0.1.0"
description = "Testbed for Catena-X boarding"
authors = ["Dénes Surman (dddenes) <surmandenes@yahoo.com>"]

[tool.poetry.dependencies]
python-dotenv = "^1.0.1"
fastapi = "0.115.7"
httpx = "^0.28.1"
jsonschema = "^4.23.0"
python = "^3.12.3"
uvicorn = "0.30.3"

[tool.poetry.dev-dependencies]
pylint = "^3.3.4"
pytest = "^8.1.1"
pytest-asyncio = "^0.26.0"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint]
max-args = 7
max-line-length = 120
max-locals=26
max-positional-arguments=7
disable = ["E1101", "W0613", "W0612", "W0621", "W0707", "W1203"]
